test: Fix tests for console exporter
